% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_p_abiss.R
\docType{data}
\name{p_abiss}
\alias{p_abiss}
\title{Annotated peptide sequences from Pyrococcus abyss}
\format{
A \code{\link[Biostrings]{AAStringSet}}
}
\usage{
p_abiss
}
\description{
The dataset contains the annotated peptide sequences from Pyrococcus abyss.
The fasta file can be downloaded from:
}
\details{
\url{https://bacteria.ensembl.org/Pyrococcus_abyssi_ge5_gca_000195935/Info/Index}
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
\keyword{datasets}
