% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_p_furiosus.R
\docType{data}
\name{p_furiosus}
\alias{p_furiosus}
\title{Annotated peptide sequences from Pyrococcus furiosus}
\format{
A \code{\link[Biostrings]{AAStringSet}}
}
\usage{
p_furiosus
}
\description{
The dataset contains the annotated peptide sequences from Pyrococcus furiosus.
The fasta file can be downloaded from:
}
\details{
<https://bacteria.ensembl.org/Pyrococcus_furiosus_dsm_3638_gca_000007305/Info/Index>
}
\author{
Robersy Sanchez (\url{https://genomaths.com}).
}
\keyword{datasets}
